+++ sample/automotive/src/main/java/com/datadog/sample/automotive/SampleAutoApplication.kt
@@ -10,7 +10,7 @@ package com.datadog.sample.automotive
 import android.app.Application
 import android.util.Log
 import com.datadog.android.Datadog
-import com.datadog.android.DatadogSite
+import com.datadog.android.FlashcatSite
 import com.datadog.android.core.configuration.BatchSize
 import com.datadog.android.core.configuration.Configuration
 import com.datadog.android.core.configuration.UploadFrequency
@@ -62,7 +62,7 @@ class SampleAutoApplication : Application() {
             env = "test",
             variant = ""
         )
-            .useSite(DatadogSite.CN)
+            .useSite(FlashcatSite.CN)
             .setBatchSize(BatchSize.SMALL)
             .setUploadFrequency(UploadFrequency.FREQUENT)
             .build()